webapp/requirements.txt
Flask==3.0.0
requests==2.31.0
gunicorn==21.2.0
python-dateutil==2.8.2
